Move CoreGroup and CxModel to builder
The CxModel was removed and merged with builder.Model. The
combined Model class is still quite small.
