+++ nengo_loihi/builder/__init__.py
@@ -1,7 +1,9 @@
 from .builder import (
     Builder,
+    CoreGroup,
     INTER_N,
     INTER_RATE,
     Model,
+    SpikeInput,
 )
 from . import connection, ensemble, node, probe
