+++ nengo_loihi/builder/node.py
@@ -7,9 +7,8 @@ from nengo_loihi.splitter import ChipReceiveNode
 @Builder.register(nengo.Node)
 def build_node(model, node):
     if isinstance(node, ChipReceiveNode):
-        cx_spiker = node.cx_spike_input
-        model.add_input(cx_spiker)
-        model.objs[node]['out'] = cx_spiker
+        model.add_input(node.spike_input)
+        model.objs[node]['out'] = node.spike_input
         return
     else:
         raise NotImplementedError()
